testing implemented

+++ src/main/java/com/cts/repository/LoanRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.cts.bean.Loan;
 
-public interface LoanRepository extends CrudRepository<Loan,Long>{
+public interface LoanRepository extends CrudRepository<Loan,Integer>{
 	
-Loan findByLoanNumber(Long loanNumber);
+Loan findByLoanNumber(int loanNumber);
 Loan findByFirstName(String firstName);
 Loan findByLastName(String lastName);
